+++ backend/src/main/java/com/example/nbe341team02/domain/product/dto/ProductDTO.java
@@ -1,4 +1,4 @@
-package com.example.nbe341team02.product.dto;
+package com.example.nbe341team02.domain.product.dto;
 
 
 import com.example.nbe341team02.product.entity.Product;
