+++ backend/src/main/java/com/example/nbe341team02/global/exception/CustomException.java
@@ -11,7 +11,6 @@ public class CustomException extends RuntimeException {
     private final String errorMessage;
 
     public CustomException(ErrorCode errorCode) {
-
         this.errorCode = errorCode;
         this.errorMessage = errorCode.getMessage();
         this.httpStatus = errorCode.getHttpStatus();
