+++ backend/src/main/java/com/example/nbe341team02/global/exception/GlobalExceptionHandler.java
@@ -1,13 +1,10 @@
-package com.example.nbe341team02.global.exception.GlobalExceptionHandler;
+package com.example.nbe341team02.global.exception;
 
-import com.example.nbe341team02.global.exception.CustomErrorResponse;
-import com.example.nbe341team02.global.exception.CustomException;
 import java.util.HashMap;
 import java.util.Map;
-import org.springframework.http.HttpStatus;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
