--- backend/src/main/java/com/example/nbe341team02/product/exception/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package com.example.nbe341team02.product.exception;
-
-public class ProductNotFoundException extends RuntimeException{
-    public ProductNotFoundException(String message) {
-        super(message);
-    }
-}
