--- backend/src/main/java/com/example/nbe341team02/product/repository/ProductRepository.java
@@ -1,10 +1,0 @@
-package com.example.nbe341team02.product.repository;
-
-import com.example.nbe341team02.product.entity.Product;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
